refactor: update types & remove references/redundant files related specifically to the ContactForm
