+++ packages/osc-ecommerce/app/components/Forms/formSchemas.ts
@@ -1,16 +1,5 @@
 import { z } from 'zod';
 
-const baseContactSchema = z.object({
-    email: z.string({ required_error: 'Email is required' }).email(),
-    firstname: z.string().trim().min(1, { message: 'Field is required' }),
-    lastname: z.string().trim().min(1, { message: 'Field is required' }),
-    phone: z.string().trim().min(1, { message: 'Field is required' }),
-});
-
-export const contactFormSchema = baseContactSchema.extend({
-    enquiry: z.string().trim().min(1, { message: 'Field is required' }),
-});
-
 export const datePickerSchema = {
     date: z.object({
         day: z
