--- packages/osc-ecommerce/app/utils/schema/zodSchema/contactFormSchema.ts
@@ -1,9 +1,0 @@
-import { z } from 'zod';
-
-export const contactFormSchema = z.object({
-    email: z.string({ required_error: 'Email is required' }).email(),
-    firstname: z.string().trim().min(1, { message: 'Field is required' }),
-    lastname: z.string().trim().min(1, { message: 'Field is required' }),
-    phone: z.string().trim().min(1, { message: 'Field is required' }),
-    enquiry: z.string().trim().min(1, { message: 'Field is required' }),
-});
