--- packages/osc-studio/schemas/objects/module/contactForm.ts
@@ -1,37 +1,0 @@
-export default {
-    name: 'form.contactForm',
-    title: 'Contact Form',
-    type: 'object',
-    groups: [
-        {
-            name: 'content',
-            title: 'Content',
-        },
-    ],
-    fields: [
-        {
-            name: 'titleAndDescription',
-            title: 'Title and Description',
-            group: 'content',
-            type: 'body',
-        },
-        {
-            name: 'termsAndConditions',
-            title: 'Terms and Conditions',
-            type: 'body',
-            group: 'content',
-        },
-        {
-            name: 'actionText',
-            title: 'Action Text',
-            description: 'Text for the submit button',
-            type: 'string',
-            group: 'content',
-        },
-    ],
-    preview: {
-        prepare() {
-            return { title: 'Contact Form' };
-        },
-    },
-};
